+++ packages/testing-tools/CHANGELOG.md
@@ -1,9 +1,9 @@
 ## 1.67.0 (2024-09-08)
 
-
 ### Features
 
-- **workspace-tools:** Added custom code for generating the Nx configuration ([651a589e](https://github.com/storm-software/storm-ops/commit/651a589e))
+- **workspace-tools:** Added custom code for generating the Nx configuration
+  ([651a589e](https://github.com/storm-software/storm-ops/commit/651a589e))
 
 ## 1.66.0 (2024-09-06)